feat: 잘못된 PR url이 아닌 경우 isFailedURL을 false로 설정
